refactor(worker): extract agent-turn Codex and xAI capacity
Move credential selection and durable capacity waits out of run.ts
into selectCodexTurnCapacity and selectXaiTurnCapacity.
